Refactoring and adding project and cce
